requirements.txt
anyio==4.3.0
asgiref==3.7.2
click==8.1.7
Django==5.0.1
djangorestframework==3.14.0
graphql-core==3.2.3
h11==0.14.0
idna==3.6
libcst==1.2.0
markdown-it-py==3.0.0
mdurl==0.1.2
mypy-extensions==1.0.0
pillow==10.2.0
Pygments==2.17.2
PyJWT==2.8.0
python-dateutil==2.8.2
python-multipart==0.0.9
pytz==2024.1
PyYAML==6.0.1
rich==13.7.0
six==1.16.0
sniffio==1.3.0
sqlparse==0.4.4
starlette==0.37.1
strawberry-graphql==0.219.2
strawberry-graphql-django==0.32.0
typer==0.9.0
typing-inspect==0.9.0
typing_extensions==4.9.0
uvicorn==0.27.1
